test: cover the class-name leak for ListBox too, not just TreeItem
Both components had the same defect and the same fix, but only TreeItem
had a regression test. Add the ListBox equivalent, and move the file to
shared/components/aria/ since it now covers className composition for two
unrelated components rather than anything tree-specific.

ListBox leaked a different set of keys (layout, orientation, state,
defaultClassName) because react-aria's render props differ per component,
so each list is asserted separately.

Both leak tests were verified red against the unfixed components. The four
pass-through tests pass in both states, which is the point: the fix
removes the leaked keys and changes nothing else.
